Move inner classes to separate folder to easier exclude controllers from generated typescript.
