+++ src/main/java/no/ndla/taxonomy/rest/v1/Relevances.java
@@ -7,17 +7,14 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.Relevance;
-import no.ndla.taxonomy.domain.Translation;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 import no.ndla.taxonomy.repositories.RelevanceRepository;
-import no.ndla.taxonomy.service.dtos.TranslationDTO;
-import no.ndla.taxonomy.service.UpdatableDto;
+import no.ndla.taxonomy.rest.v1.dtos.RelevanceDTO;
+import no.ndla.taxonomy.rest.v1.dtos.RelevancePUT;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -26,9 +23,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @RestController
@@ -80,56 +74,4 @@ public class Relevances extends CrudController<Relevance> {
         updateEntity(id, command);
     }
 
-    @Schema(name = "Relevance")
-    public static class RelevanceDTO {
-        @JsonProperty
-        @Schema(example = "urn:relevance:core")
-        public URI id;
-
-        @JsonProperty
-        @Schema(description = "The name of the relevance", example = "Core")
-        public String name;
-
-        @JsonProperty
-        @Schema(description = "All translations of this relevance")
-        private Set<TranslationDTO> translations;
-
-        @JsonProperty
-        @Schema(description = "List of language codes supported by translations")
-        private Set<String> supportedLanguages;
-
-        public RelevanceDTO() {
-        }
-
-        public RelevanceDTO(Relevance relevance, String language) {
-            this.id = relevance.getPublicId();
-
-            var translations = relevance.getTranslations();
-            this.translations = translations.stream().map(TranslationDTO::new).collect(Collectors.toSet());
-            this.supportedLanguages = this.translations.stream().map(t -> t.language).collect(Collectors.toSet());
-
-            this.name = translations.stream().filter(t -> Objects.equals(t.getLanguageCode(), language)).findFirst()
-                    .map(Translation::getName).orElse(relevance.getName());
-        }
-    }
-
-    public static class RelevancePUT implements UpdatableDto<Relevance> {
-        @JsonProperty
-        @Schema(description = "If specified, set the id to this value. Must start with urn:relevance: and be a valid URI. If ommitted, an id will be assigned automatically. Ignored on update", example = "urn:relevance:supplementary")
-        public URI id;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The name of the relevance", example = "Supplementary")
-        public String name;
-
-        @Override
-        public Optional<URI> getId() {
-            return Optional.ofNullable(id);
-        }
-
-        @Override
-        public void apply(Relevance entity) {
-            entity.setName(name);
-        }
-    }
 }
