+++ src/main/java/no/ndla/taxonomy/rest/v1/ResourceResourceTypes.java
@@ -7,16 +7,16 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.ResourceResourceType;
 import no.ndla.taxonomy.domain.ResourceType;
 import no.ndla.taxonomy.repositories.NodeRepository;
 import no.ndla.taxonomy.repositories.ResourceResourceTypeRepository;
 import no.ndla.taxonomy.repositories.ResourceTypeRepository;
+import no.ndla.taxonomy.rest.v1.dtos.ResourceResourceTypeDTO;
+import no.ndla.taxonomy.rest.v1.dtos.ResourceResourceTypePOST;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -86,37 +86,4 @@ public class ResourceResourceTypes {
         return new ResourceResourceTypeDTO(result);
     }
 
-    public static class ResourceResourceTypePOST {
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Resource id", example = "urn:resource:123")
-        URI resourceId;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Resource type id", example = "urn:resourcetype:234")
-        URI resourceTypeId;
-    }
-
-    @Schema(name = "ResourceResourceType")
-    public static class ResourceResourceTypeDTO {
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Resource type id", example = "urn:resource:123")
-        URI resourceId;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Resource type id", example = "urn:resourcetype:234")
-        URI resourceTypeId;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Resource to resource type connection id", example = "urn:resource-has-resourcetypes:12")
-        URI id;
-
-        public ResourceResourceTypeDTO() {
-        }
-
-        public ResourceResourceTypeDTO(ResourceResourceType resourceResourceType) {
-            id = resourceResourceType.getPublicId();
-            resourceId = resourceResourceType.getNode().getPublicId();
-            resourceTypeId = resourceResourceType.getResourceType().getPublicId();
-        }
-    }
 }
