+++ src/main/java/no/ndla/taxonomy/rest/v1/ResourceTranslations.java
@@ -7,13 +7,12 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 import no.ndla.taxonomy.repositories.NodeRepository;
+import no.ndla.taxonomy.rest.v1.dtos.TranslationPUT;
 import no.ndla.taxonomy.service.dtos.TranslationDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -80,7 +79,7 @@ public class ResourceTranslations {
     @Transactional
     public void createUpdateResourceTranslation(@PathVariable("id") URI id,
             @Parameter(description = "ISO-639-1 language code", example = "nb", required = true) @PathVariable("language") String language,
-            @Parameter(name = "resource", description = "The new or updated translation") @RequestBody ResourceTranslationPUT command) {
+            @Parameter(name = "resource", description = "The new or updated translation") @RequestBody TranslationPUT command) {
         var resource = nodeRepository.getByPublicId(id);
         resource.addTranslation(command.name, language);
         entityManager.persist(resource);
@@ -101,9 +100,4 @@ public class ResourceTranslations {
         });
     }
 
-    public static class ResourceTranslationPUT {
-        @JsonProperty
-        @Schema(description = "The translated name of the resource", example = "Introduction to algebra")
-        public String name;
-    }
 }
