+++ src/main/java/no/ndla/taxonomy/rest/v1/ResourceTypes.java
@@ -7,18 +7,14 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import no.ndla.taxonomy.domain.JsonTranslation;
 import no.ndla.taxonomy.domain.ResourceType;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 import no.ndla.taxonomy.repositories.ResourceTypeRepository;
-import no.ndla.taxonomy.service.dtos.TranslationDTO;
-import no.ndla.taxonomy.service.UpdatableDto;
+import no.ndla.taxonomy.rest.v1.dtos.ResourceTypeDTO;
+import no.ndla.taxonomy.rest.v1.dtos.ResourceTypePUT;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -26,10 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @RestController
@@ -108,71 +101,4 @@ public class ResourceTypes extends CrudController<ResourceType> {
                 .map(resourceType -> new ResourceTypeDTO(resourceType, language, 100)).collect(Collectors.toList());
     }
 
-    @Schema(name = "ResourceType")
-    public static class ResourceTypeDTO {
-        @JsonProperty
-        @Schema(example = "urn:resourcetype:1")
-        public URI id;
-
-        @JsonProperty
-        @Schema(description = "The name of the resource type", example = "Lecture")
-        public String name;
-
-        @JsonProperty
-        @Schema(description = "Sub resource types")
-        @JsonInclude(JsonInclude.Include.NON_EMPTY)
-        public List<ResourceTypeDTO> subtypes = new ArrayList<>();
-
-        @JsonProperty
-        @Schema(description = "All translations of this resource type")
-        private Set<TranslationDTO> translations;
-
-        @JsonProperty
-        @Schema(description = "List of language codes supported by translations")
-        private Set<String> supportedLanguages;
-
-        public ResourceTypeDTO() {
-        }
-
-        public ResourceTypeDTO(ResourceType resourceType, String language, int recursionLevels) {
-            this.id = resourceType.getPublicId();
-
-            var translations = resourceType.getTranslations();
-            this.translations = translations.stream().map(TranslationDTO::new).collect(Collectors.toSet());
-            this.supportedLanguages = this.translations.stream().map(t -> t.language).collect(Collectors.toSet());
-
-            this.name = resourceType.getTranslation(language).map(JsonTranslation::getName)
-                    .orElse(resourceType.getName());
-
-            if (recursionLevels > 0) {
-                this.subtypes = resourceType.getSubtypes().stream()
-                        .map(resourceType1 -> new ResourceTypeDTO(resourceType1, language, recursionLevels - 1))
-                        .collect(Collectors.toList());
-            }
-        }
-    }
-
-    public static class ResourceTypePUT implements UpdatableDto<ResourceType> {
-        @JsonProperty
-        @Schema(description = "If specified, the new resource type will be a child of the mentioned resource type.")
-        public URI parentId;
-
-        @JsonProperty
-        @Schema(description = "If specified, set the id to this value. Must start with urn:resourcetype: and be a valid URI. If omitted, an id will be assigned automatically.", example = "urn:resourcetype:1")
-        public URI id;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The name of the resource type", example = "Lecture")
-        public String name;
-
-        @Override
-        public Optional<URI> getId() {
-            return Optional.ofNullable(id);
-        }
-
-        @Override
-        public void apply(ResourceType entity) {
-            entity.setName(name);
-        }
-    }
 }
