+++ src/main/java/no/ndla/taxonomy/rest/v1/TopicSubtopics.java
@@ -7,10 +7,8 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeConnection;
@@ -18,6 +16,9 @@ import no.ndla.taxonomy.domain.Relevance;
 import no.ndla.taxonomy.repositories.NodeConnectionRepository;
 import no.ndla.taxonomy.repositories.NodeRepository;
 import no.ndla.taxonomy.repositories.RelevanceRepository;
+import no.ndla.taxonomy.rest.v1.dtos.TopicSubtopicDTO;
+import no.ndla.taxonomy.rest.v1.dtos.TopicSubtopicPOST;
+import no.ndla.taxonomy.rest.v1.dtos.TopicSubtopicPUT;
 import no.ndla.taxonomy.service.NodeConnectionService;
 import no.ndla.taxonomy.service.dtos.SearchResultDTO;
 import org.springframework.data.domain.PageRequest;
@@ -143,82 +144,4 @@ public class TopicSubtopics {
         connectionService.updateParentChild(topicSubtopic, relevance, rank, Optional.empty());
     }
 
-    public static class TopicSubtopicPOST {
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Topic id", example = "urn:topic:234")
-        public URI topicid;
-
-        @JsonProperty
-        @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "Subtopic id", example = "urn:topic:234")
-        public URI subtopicid;
-
-        @JsonProperty
-        @Schema(description = "Backwards compatibility: Always true. Ignored on insert/update", example = "true")
-        public boolean primary = true;
-
-        @JsonProperty
-        @Schema(description = "Order in which to sort the subtopic for the topic", example = "1")
-        public int rank;
-
-        @JsonProperty
-        @Schema(description = "Relevance id", example = "urn:relevance:core")
-        public URI relevanceId;
-    }
-
-    public static class TopicSubtopicPUT {
-        @JsonProperty
-        @Schema(description = "Connection id", example = "urn:topic-has-subtopics:345")
-        public URI id;
-
-        @JsonProperty
-        @Schema(description = "Backwards compatibility: Always true. Ignored on insert/update", example = "true")
-        public boolean primary;
-
-        @JsonProperty
-        @Schema(description = "Order in which subtopic is sorted for the topic", example = "1")
-        public int rank;
-
-        @JsonProperty
-        @Schema(description = "Relevance id", example = "urn:relevance:core")
-        public URI relevanceId;
-    }
-
-    @Schema(name = "TopicSubtopic")
-    public static class TopicSubtopicDTO {
-        @JsonProperty
-        @Schema(description = "Topic id", example = "urn:topic:234")
-        public URI topicid;
-
-        @JsonProperty
-        @Schema(description = "Subtopic id", example = "urn:topic:234")
-        public URI subtopicid;
-
-        @JsonProperty
-        @Schema(description = "Connection id", example = "urn:topic-has-subtopics:345")
-        public URI id;
-
-        @JsonProperty
-        @Schema(description = "Backwards compatibility: Always true. Ignored on insert/update", example = "true")
-        public boolean primary;
-
-        @JsonProperty
-        @Schema(description = "Order in which subtopic is sorted for the topic", example = "1")
-        public int rank;
-
-        @JsonProperty
-        @Schema(description = "Relevance id", example = "urn:relevance:core")
-        public URI relevanceId;
-
-        TopicSubtopicDTO() {
-        }
-
-        TopicSubtopicDTO(NodeConnection nodeConnection) {
-            id = nodeConnection.getPublicId();
-            nodeConnection.getParent().ifPresent(topic -> topicid = topic.getPublicId());
-            nodeConnection.getChild().ifPresent(subtopic -> subtopicid = subtopic.getPublicId());
-            relevanceId = nodeConnection.getRelevance().map(Relevance::getPublicId).orElse(null);
-            primary = true;
-            rank = nodeConnection.getRank();
-        }
-    }
 }
