+++ src/main/java/no/ndla/taxonomy/rest/v1/TopicTranslations.java
@@ -7,14 +7,13 @@
 
 package no.ndla.taxonomy.rest.v1;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 import no.ndla.taxonomy.repositories.NodeRepository;
+import no.ndla.taxonomy.rest.v1.dtos.TranslationPUT;
 import no.ndla.taxonomy.service.dtos.TranslationDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -81,7 +80,7 @@ public class TopicTranslations {
     @Transactional
     public void createUpdateTopicTranslation(@PathVariable("id") URI id,
             @Parameter(description = "ISO-639-1 language code", example = "nb", required = true) @PathVariable("language") String language,
-            @Parameter(name = "topic", description = "The new or updated translation") @RequestBody TopicTranslationPUT command) {
+            @Parameter(name = "topic", description = "The new or updated translation") @RequestBody TranslationPUT command) {
         var topic = nodeRepository.getByPublicId(id);
         topic.addTranslation(command.name, language);
         entityManager.persist(topic);
@@ -102,9 +101,4 @@ public class TopicTranslations {
         });
     }
 
-    public static class TopicTranslationPUT {
-        @JsonProperty
-        @Schema(description = "The translated name of the topic", example = "Trigonometry")
-        public String name;
-    }
 }
