+++ src/main/java/no/ndla/taxonomy/rest/v1/dtos/ContextPOST.java
@@ -1,0 +1,14 @@
+/*
+ * Part of NDLA taxonomy-api
+ * Copyright (C) 2023 NDLA
+ *
+ * See LICENSE
+ */
+
+package no.ndla.taxonomy.rest.v1.dtos;
+
+import java.net.URI;
+
+public class ContextPOST {
+    public URI id;
+}
