+++ src/test/java/no/ndla/taxonomy/rest/v1/NodeResourcesTest.java
@@ -12,6 +12,8 @@ import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeConnection;
 import no.ndla.taxonomy.domain.NodeType;
 import no.ndla.taxonomy.rest.v1.dtos.NodeResourceDTO;
+import no.ndla.taxonomy.rest.v1.dtos.NodeResourcePOST;
+import no.ndla.taxonomy.rest.v1.dtos.NodeResourcePUT;
 import no.ndla.taxonomy.service.dtos.MetadataDto;
 import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import org.junit.jupiter.api.Test;
@@ -34,7 +36,7 @@ public class NodeResourcesTest extends RestTest {
         resource.setName("Introduction to integration");
         integrationId = resource.getPublicId();
 
-        URI id = getId(testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        URI id = getId(testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 nodeId = calculusId;
                 resourceId = integrationId;
@@ -67,7 +69,7 @@ public class NodeResourcesTest extends RestTest {
         // parameter and will be
         // forced to become primary (subject-topics already did this)
 
-        final var id = getId(testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        final var id = getId(testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 nodeId = calculusId;
                 resourceId = integrationId;
@@ -85,7 +87,7 @@ public class NodeResourcesTest extends RestTest {
         // After behavior change: Add the resource again to another topic with primary = false
         // should create a non-primary resource connection
         final var resource2ConnectionPublicId = getId(
-                testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+                testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
                     {
                         nodeId = node2Id;
                         resourceId = integrationId;
@@ -115,7 +117,7 @@ public class NodeResourcesTest extends RestTest {
         final var calculusId = calculus.getPublicId();
         final var integrationId = integration.getPublicId();
 
-        testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 nodeId = calculusId;
                 resourceId = integrationId;
@@ -137,7 +139,7 @@ public class NodeResourcesTest extends RestTest {
     public void can_update_node_resource() throws Exception {
         URI id = save(NodeConnection.create(newTopic(), newResource())).getPublicId();
 
-        testUtils.updateResource("/v1/node-resources/" + id, new NodeResources.NodeResourcePUT() {
+        testUtils.updateResource("/v1/node-resources/" + id, new NodeResourcePUT() {
             {
                 primary = true;
             }
@@ -150,7 +152,7 @@ public class NodeResourcesTest extends RestTest {
     public void cannot_unset_primary_node() throws Exception {
         URI id = save(NodeConnection.create(newTopic(), newResource(), true)).getPublicId();
 
-        testUtils.updateResource("/v1/node-resources/" + id, new NodeResources.NodeResourcePUT() {
+        testUtils.updateResource("/v1/node-resources/" + id, new NodeResourcePUT() {
             {
                 primary = false;
             }
@@ -221,7 +223,7 @@ public class NodeResourcesTest extends RestTest {
 
         Node graphTheory = builder.node(NodeType.TOPIC, t -> t.name("graph theory"));
 
-        testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 nodeId = graphTheory.getPublicId();
                 resourceId = graphs.getPublicId();
@@ -246,14 +248,14 @@ public class NodeResourcesTest extends RestTest {
 
         URI geometrySquares = save(NodeConnection.create(geometry, squares)).getPublicId();
         URI geometryCircles = save(NodeConnection.create(geometry, circles)).getPublicId();
-        testUtils.updateResource("/v1/node-resources/" + geometryCircles, new NodeResources.NodeResourcePUT() {
+        testUtils.updateResource("/v1/node-resources/" + geometryCircles, new NodeResourcePUT() {
             {
                 primary = true;
                 id = geometryCircles;
                 rank = 1;
             }
         });
-        testUtils.updateResource("/v1/node-resources/" + geometrySquares, new NodeResources.NodeResourcePUT() {
+        testUtils.updateResource("/v1/node-resources/" + geometrySquares, new NodeResourcePUT() {
             {
                 primary = true;
                 id = geometrySquares;
@@ -283,7 +285,7 @@ public class NodeResourcesTest extends RestTest {
         var squares = builder.node(NodeType.RESOURCE, r -> r.name("Squares").publicId("urn:resource:1"));
         var circles = builder.node(NodeType.RESOURCE, r -> r.name("Circles").publicId("urn:resource:2"));
 
-        testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 primary = true;
                 nodeId = geometry.getPublicId();
@@ -292,7 +294,7 @@ public class NodeResourcesTest extends RestTest {
             }
         });
 
-        testUtils.createResource("/v1/node-resources", new NodeResources.NodeResourcePOST() {
+        testUtils.createResource("/v1/node-resources", new NodeResourcePOST() {
             {
                 primary = true;
                 nodeId = geometry.getPublicId();
@@ -318,7 +320,7 @@ public class NodeResourcesTest extends RestTest {
         var updatedConnection = nodeResources.get(nodeResources.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/node-resources/" + updatedConnection.getPublicId().toString(),
-                new NodeResources.NodeResourcePUT() {
+                new NodeResourcePUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -349,7 +351,7 @@ public class NodeResourcesTest extends RestTest {
         var updatedConnection = nodeResources.get(nodeResources.size() - 1);
         assertEquals(100, updatedConnection.getRank());
         testUtils.updateResource("/v1/node-resources/" + updatedConnection.getPublicId().toString(),
-                new NodeResources.NodeResourcePUT() {
+                new NodeResourcePUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -382,7 +384,7 @@ public class NodeResourcesTest extends RestTest {
         var updatedConnection = nodeResources.get(nodeResources.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/node-resources/" + nodeResources.get(9).getPublicId().toString(),
-                new NodeResources.NodeResourcePUT() {
+                new NodeResourcePUT() {
                     {
                         primary = true;
                         rank = 99;
