+++ src/test/java/no/ndla/taxonomy/rest/v1/ResourceTypeTranslationsTest.java
@@ -8,6 +8,8 @@
 package no.ndla.taxonomy.rest.v1;
 
 import no.ndla.taxonomy.domain.ResourceType;
+import no.ndla.taxonomy.rest.v1.dtos.ResourceTypeDTO;
+import no.ndla.taxonomy.rest.v1.dtos.TranslationPUT;
 import no.ndla.taxonomy.service.dtos.TranslationDTO;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletResponse;
@@ -27,8 +29,7 @@ public class ResourceTypeTranslationsTest extends RestTest {
         builder.resourceType(t -> t.name("Lecture").translation("Forelesning", "nb"));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/resource-types?language=nb");
-        ResourceTypes.ResourceTypeDTO[] resourceTypes = testUtils.getObject(ResourceTypes.ResourceTypeDTO[].class,
-                response);
+        ResourceTypeDTO[] resourceTypes = testUtils.getObject(ResourceTypeDTO[].class, response);
 
         assertEquals(2, resourceTypes.length);
         assertAnyTrue(resourceTypes, s -> s.name.equals("Artikkel"));
@@ -39,14 +40,14 @@ public class ResourceTypeTranslationsTest extends RestTest {
     public void can_get_single_resource_type() throws Exception {
         URI id = builder.resourceType(t -> t.name("Article").translation("Artikkel", "nb")).getPublicId();
 
-        ResourceTypes.ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, "nb");
+        ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, "nb");
         assertEquals("Artikkel", resourceType.name);
     }
 
     @Test
     public void fallback_to_default_language() throws Exception {
         URI id = builder.resourceType(t -> t.name("Article")).getPublicId();
-        ResourceTypes.ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, "XX");
+        ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, "XX");
         assertEquals("Article", resourceType.name);
     }
 
@@ -54,7 +55,7 @@ public class ResourceTypeTranslationsTest extends RestTest {
     public void can_get_default_language() throws Exception {
         URI id = builder.resourceType(t -> t.name("Article").translation("Artikkel", "nb")).getPublicId();
 
-        ResourceTypes.ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, null);
+        ResourceTypeDTO resourceType = getResourceTypeIndexDocument(id, null);
         assertEquals("Article", resourceType.name);
     }
 
@@ -63,12 +64,11 @@ public class ResourceTypeTranslationsTest extends RestTest {
         ResourceType article = builder.resourceType(t -> t.name("Article"));
         URI id = article.getPublicId();
 
-        testUtils.updateResource("/v1/resource-types/" + id + "/translations/nb",
-                new ResourceTypeTranslations.ResourceTypeTranslationPUT() {
-                    {
-                        name = "Artikkel";
-                    }
-                });
+        testUtils.updateResource("/v1/resource-types/" + id + "/translations/nb", new TranslationPUT() {
+            {
+                name = "Artikkel";
+            }
+        });
 
         assertEquals("Artikkel", article.getTranslation("nb").get().getName());
     }
@@ -109,10 +109,10 @@ public class ResourceTypeTranslationsTest extends RestTest {
         assertEquals("nb", translation.language);
     }
 
-    private ResourceTypes.ResourceTypeDTO getResourceTypeIndexDocument(URI id, String language) throws Exception {
+    private ResourceTypeDTO getResourceTypeIndexDocument(URI id, String language) throws Exception {
         String path = "/v1/resource-types/" + id;
         if (isNotEmpty(language))
             path = path + "?language=" + language;
-        return testUtils.getObject(ResourceTypes.ResourceTypeDTO.class, testUtils.getResource(path));
+        return testUtils.getObject(ResourceTypeDTO.class, testUtils.getResource(path));
     }
 }
