+++ src/test/java/no/ndla/taxonomy/rest/v1/SubjectTopicsTest.java
@@ -10,6 +10,9 @@ package no.ndla.taxonomy.rest.v1;
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeConnection;
 import no.ndla.taxonomy.domain.NodeType;
+import no.ndla.taxonomy.rest.v1.dtos.SubjectTopicDTO;
+import no.ndla.taxonomy.rest.v1.dtos.SubjectTopicPOST;
+import no.ndla.taxonomy.rest.v1.dtos.SubjectTopicPUT;
 import no.ndla.taxonomy.service.RankableConnectionUpdater;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -39,7 +42,7 @@ public class SubjectTopicsTest extends RestTest {
         subjectId = newSubject().name("physics").getPublicId();
         topicId = newTopic().name("trigonometry").getPublicId();
 
-        URI id = getId(testUtils.createResource("/v1/subject-topics", new SubjectTopics.SubjectTopicPOST() {
+        URI id = getId(testUtils.createResource("/v1/subject-topics", new SubjectTopicPOST() {
             {
                 this.subjectid = subjectId;
                 this.topicid = topicId;
@@ -63,7 +66,7 @@ public class SubjectTopicsTest extends RestTest {
         URI subjectId = physics.getPublicId();
         URI topicId = trigonometry.getPublicId();
 
-        testUtils.createResource("/v1/subject-topics", new SubjectTopics.SubjectTopicPOST() {
+        testUtils.createResource("/v1/subject-topics", new SubjectTopicPOST() {
             {
                 this.subjectid = subjectId;
                 this.topicid = topicId;
@@ -83,11 +86,10 @@ public class SubjectTopicsTest extends RestTest {
         URI id = save(NodeConnection.create(newSubject(), newTopic())).getPublicId();
 
         MockHttpServletResponse responseBefore = testUtils.getResource("/v1/subject-topics/" + id.toString());
-        SubjectTopics.SubjectTopicDTO connection = testUtils.getObject(SubjectTopics.SubjectTopicDTO.class,
-                responseBefore);
+        SubjectTopicDTO connection = testUtils.getObject(SubjectTopicDTO.class, responseBefore);
         assertEquals(0, connection.rank);
 
-        testUtils.updateResource("/v1/subject-topics/" + id, new SubjectTopics.SubjectTopicPUT() {
+        testUtils.updateResource("/v1/subject-topics/" + id, new SubjectTopicPUT() {
             {
                 primary = true;
                 rank = 12;
@@ -95,8 +97,7 @@ public class SubjectTopicsTest extends RestTest {
         });
 
         MockHttpServletResponse responseAfter = testUtils.getResource("/v1/subject-topics/" + id);
-        SubjectTopics.SubjectTopicDTO connectionAfter = testUtils.getObject(SubjectTopics.SubjectTopicDTO.class,
-                responseAfter);
+        SubjectTopicDTO connectionAfter = testUtils.getObject(SubjectTopicDTO.class, responseAfter);
 
         assertEquals(12, connectionAfter.rank);
     }
@@ -111,7 +112,7 @@ public class SubjectTopicsTest extends RestTest {
         NodeConnection updatedConnection = subjectTopics.get(subjectTopics.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/subject-topics/" + updatedConnection.getPublicId().toString(),
-                new SubjectTopics.SubjectTopicPUT() {
+                new SubjectTopicPUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -123,8 +124,7 @@ public class SubjectTopicsTest extends RestTest {
         for (NodeConnection subjectTopic : subjectTopics) {
             MockHttpServletResponse response = testUtils
                     .getResource("/v1/subject-topics/" + subjectTopic.getPublicId().toString());
-            SubjectTopics.SubjectTopicDTO connectionFromDb = testUtils.getObject(SubjectTopics.SubjectTopicDTO.class,
-                    response);
+            SubjectTopicDTO connectionFromDb = testUtils.getObject(SubjectTopicDTO.class, response);
             // verify that the other connections have had their rank bumped up 1
             if (!connectionFromDb.id.equals(updatedConnection.getPublicId())) {
                 int oldRank = mappedRanks.get(connectionFromDb.id.toString());
@@ -143,7 +143,7 @@ public class SubjectTopicsTest extends RestTest {
         NodeConnection updatedConnection = subjectTopics.get(subjectTopics.size() - 1);
         assertEquals(100, updatedConnection.getRank());
         testUtils.updateResource("/v1/subject-topics/" + updatedConnection.getPublicId().toString(),
-                new SubjectTopics.SubjectTopicPUT() {
+                new SubjectTopicPUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -155,8 +155,7 @@ public class SubjectTopicsTest extends RestTest {
         for (NodeConnection subjectTopic : subjectTopics) {
             MockHttpServletResponse response = testUtils
                     .getResource("/v1/subject-topics/" + subjectTopic.getPublicId().toString());
-            SubjectTopics.SubjectTopicDTO connectionFromDb = testUtils.getObject(SubjectTopics.SubjectTopicDTO.class,
-                    response);
+            SubjectTopicDTO connectionFromDb = testUtils.getObject(SubjectTopicDTO.class, response);
             // verify that only the contiguous connections are updated
             if (!connectionFromDb.id.equals(updatedConnection.getPublicId())) {
                 int oldRank = mappedRanks.get(connectionFromDb.id.toString());
@@ -178,7 +177,7 @@ public class SubjectTopicsTest extends RestTest {
         NodeConnection updatedConnection = subjectTopics.get(subjectTopics.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/subject-topics/" + subjectTopics.get(9).getPublicId().toString(),
-                new SubjectTopics.SubjectTopicPUT() {
+                new SubjectTopicPUT() {
                     {
                         primary = true;
                         rank = 99;
@@ -190,8 +189,7 @@ public class SubjectTopicsTest extends RestTest {
         for (NodeConnection subjectTopic : subjectTopics) {
             MockHttpServletResponse response = testUtils
                     .getResource("/v1/subject-topics/" + subjectTopic.getPublicId().toString());
-            SubjectTopics.SubjectTopicDTO connection = testUtils.getObject(SubjectTopics.SubjectTopicDTO.class,
-                    response);
+            SubjectTopicDTO connection = testUtils.getObject(SubjectTopicDTO.class, response);
             if (!connection.id.equals(updatedConnection.getPublicId())) {
                 assertEquals(mappedRanks.get(connection.id.toString()).intValue(), connection.rank);
             }
@@ -224,8 +222,7 @@ public class SubjectTopicsTest extends RestTest {
         URI trigonometryId = trigonometry.getPublicId();
 
         MockHttpServletResponse response = testUtils.getResource("/v1/subject-topics");
-        SubjectTopics.SubjectTopicDTO[] subjectTopics = testUtils.getObject(SubjectTopics.SubjectTopicDTO[].class,
-                response);
+        SubjectTopicDTO[] subjectTopics = testUtils.getObject(SubjectTopicDTO[].class, response);
 
         assertEquals(2, subjectTopics.length);
         assertAnyTrue(subjectTopics, t -> physicsId.equals(t.subjectid) && electricityId.equals(t.topicid));
@@ -244,8 +241,7 @@ public class SubjectTopicsTest extends RestTest {
         URI id = subjectTopic.getPublicId();
 
         MockHttpServletResponse resource = testUtils.getResource("/v1/subject-topics/" + id);
-        SubjectTopics.SubjectTopicDTO subjectTopicIndexDocument = testUtils
-                .getObject(SubjectTopics.SubjectTopicDTO.class, resource);
+        SubjectTopicDTO subjectTopicIndexDocument = testUtils.getObject(SubjectTopicDTO.class, resource);
         assertEquals(subjectid, subjectTopicIndexDocument.subjectid);
         assertEquals(topicid, subjectTopicIndexDocument.topicid);
     }
@@ -256,7 +252,7 @@ public class SubjectTopicsTest extends RestTest {
                 s -> s.isContext(true).name("Mathematics").child(NodeType.TOPIC, t -> t.name("Geometry")));
 
         MockHttpServletResponse response = testUtils.getResource("/v1/subject-topics");
-        SubjectTopics.SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopics.SubjectTopicDTO[].class, response);
+        SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopicDTO[].class, response);
 
         assertAllTrue(topics, t -> t.rank == 0);
     }
@@ -270,26 +266,24 @@ public class SubjectTopicsTest extends RestTest {
         NodeConnection geometryMaths = save(NodeConnection.create(mathematics, geometry));
         NodeConnection statisticsMaths = save(NodeConnection.create(mathematics, statistics));
 
-        testUtils.updateResource("/v1/subject-topics/" + geometryMaths.getPublicId(),
-                new SubjectTopics.SubjectTopicPUT() {
-                    {
-                        primary = true;
-                        id = geometryMaths.getPublicId();
-                        rank = 2;
-                    }
-                });
+        testUtils.updateResource("/v1/subject-topics/" + geometryMaths.getPublicId(), new SubjectTopicPUT() {
+            {
+                primary = true;
+                id = geometryMaths.getPublicId();
+                rank = 2;
+            }
+        });
 
-        testUtils.updateResource("/v1/subject-topics/" + statisticsMaths.getPublicId(),
-                new SubjectTopics.SubjectTopicPUT() {
-                    {
-                        primary = true;
-                        id = statisticsMaths.getPublicId();
-                        rank = 1;
-                    }
-                });
+        testUtils.updateResource("/v1/subject-topics/" + statisticsMaths.getPublicId(), new SubjectTopicPUT() {
+            {
+                primary = true;
+                id = statisticsMaths.getPublicId();
+                rank = 1;
+            }
+        });
 
         MockHttpServletResponse response = testUtils.getResource("/v1/subjects/urn:subject:1/topics");
-        SubjectTopics.SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopics.SubjectTopicDTO[].class, response);
+        SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopicDTO[].class, response);
 
         assertEquals(statistics.getPublicId(), topics[0].id);
         assertEquals(geometry.getPublicId(), topics[1].id);
@@ -308,39 +302,37 @@ public class SubjectTopicsTest extends RestTest {
         NodeConnection tst1 = save(NodeConnection.create(geometry, subtopic1));
         NodeConnection tst2 = save(NodeConnection.create(geometry, subtopic2));
 
-        testUtils.updateResource("/v1/subject-topics/" + geometryMaths.getPublicId(),
-                new SubjectTopics.SubjectTopicPUT() {
-                    {
-                        primary = true;
-                        id = geometryMaths.getPublicId();
-                        rank = 2;
-                    }
-                });
+        testUtils.updateResource("/v1/subject-topics/" + geometryMaths.getPublicId(), new SubjectTopicPUT() {
+            {
+                primary = true;
+                id = geometryMaths.getPublicId();
+                rank = 2;
+            }
+        });
 
-        testUtils.updateResource("/v1/subject-topics/" + statisticsMaths.getPublicId(),
-                new SubjectTopics.SubjectTopicPUT() {
-                    {
-                        primary = true;
-                        id = statisticsMaths.getPublicId();
-                        rank = 1;
-                    }
-                });
+        testUtils.updateResource("/v1/subject-topics/" + statisticsMaths.getPublicId(), new SubjectTopicPUT() {
+            {
+                primary = true;
+                id = statisticsMaths.getPublicId();
+                rank = 1;
+            }
+        });
 
-        testUtils.updateResource("/v1/topic-subtopics/" + tst1.getPublicId(), new SubjectTopics.SubjectTopicPUT() {
+        testUtils.updateResource("/v1/topic-subtopics/" + tst1.getPublicId(), new SubjectTopicPUT() {
             {
                 primary = true;
                 rank = 2;
             }
         });
 
-        testUtils.updateResource("/v1/topic-subtopics/" + tst2.getPublicId(), new SubjectTopics.SubjectTopicPUT() {
+        testUtils.updateResource("/v1/topic-subtopics/" + tst2.getPublicId(), new SubjectTopicPUT() {
             {
                 primary = true;
                 rank = 1;
             }
         });
         MockHttpServletResponse response = testUtils.getResource("/v1/subjects/urn:subject:1/topics?recursive=true");
-        SubjectTopics.SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopics.SubjectTopicDTO[].class, response);
+        SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopicDTO[].class, response);
 
         assertEquals(statistics.getPublicId(), topics[0].id);
         assertEquals(geometry.getPublicId(), topics[1].id);
@@ -355,14 +347,14 @@ public class SubjectTopicsTest extends RestTest {
         Node geometry = builder.node(NodeType.TOPIC, t -> t.name("Geometry").publicId("urn:topic:1"));
         Node statistics = builder.node(NodeType.TOPIC, t -> t.name("Statistics").publicId("urn:topic:2"));
 
-        testUtils.createResource("/v1/subject-topics", new SubjectTopics.SubjectTopicPOST() {
+        testUtils.createResource("/v1/subject-topics", new SubjectTopicPOST() {
             {
                 subjectid = mathematics.getPublicId();
                 topicid = geometry.getPublicId();
                 rank = 2;
             }
         });
-        testUtils.createResource("/v1/subject-topics", new SubjectTopics.SubjectTopicPOST() {
+        testUtils.createResource("/v1/subject-topics", new SubjectTopicPOST() {
             {
                 subjectid = mathematics.getPublicId();
                 topicid = statistics.getPublicId();
@@ -371,7 +363,7 @@ public class SubjectTopicsTest extends RestTest {
         });
 
         MockHttpServletResponse response = testUtils.getResource("/v1/subjects/urn:subject:1/topics");
-        SubjectTopics.SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopics.SubjectTopicDTO[].class, response);
+        SubjectTopicDTO[] topics = testUtils.getObject(SubjectTopicDTO[].class, response);
 
         assertEquals(statistics.getPublicId(), topics[0].id);
         assertEquals(geometry.getPublicId(), topics[1].id);
