+++ src/test/java/no/ndla/taxonomy/rest/v1/SubjectTranslationsTest.java
@@ -9,6 +9,7 @@ package no.ndla.taxonomy.rest.v1;
 
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeType;
+import no.ndla.taxonomy.rest.v1.dtos.TranslationPUT;
 import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import no.ndla.taxonomy.service.dtos.NodeDTO;
 import no.ndla.taxonomy.service.dtos.TranslationDTO;
@@ -72,12 +73,11 @@ public class SubjectTranslationsTest extends RestTest {
         Node mathematics = builder.node(NodeType.SUBJECT, s -> s.name("Mathematics"));
         URI id = mathematics.getPublicId();
 
-        testUtils.updateResource("/v1/subjects/" + id + "/translations/nb",
-                new SubjectTranslations.SubjectTranslationPUT() {
-                    {
-                        name = "Matematikk";
-                    }
-                });
+        testUtils.updateResource("/v1/subjects/" + id + "/translations/nb", new TranslationPUT() {
+            {
+                name = "Matematikk";
+            }
+        });
 
         assertEquals("Matematikk", mathematics.getTranslation("nb").get().getName());
     }
