+++ src/test/java/no/ndla/taxonomy/rest/v1/TopicResourcesTest.java
@@ -10,6 +10,9 @@ package no.ndla.taxonomy.rest.v1;
 import no.ndla.taxonomy.domain.Node;
 import no.ndla.taxonomy.domain.NodeConnection;
 import no.ndla.taxonomy.domain.NodeType;
+import no.ndla.taxonomy.rest.v1.dtos.TopicResourceDTO;
+import no.ndla.taxonomy.rest.v1.dtos.TopicResourcePOST;
+import no.ndla.taxonomy.rest.v1.dtos.TopicResourcePUT;
 import no.ndla.taxonomy.service.dtos.NodeChildDTO;
 import org.junit.jupiter.api.Test;
 
@@ -33,7 +36,7 @@ public class TopicResourcesTest extends RestTest {
         resource.setName("Introduction to integration");
         integrationId = resource.getPublicId();
 
-        URI id = getId(testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        URI id = getId(testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 topicid = calculusId;
                 resourceId = integrationId;
@@ -66,7 +69,7 @@ public class TopicResourcesTest extends RestTest {
         // parameter and will be
         // forced to become primary (subject-topics already did this)
 
-        final var id = getId(testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        final var id = getId(testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 topicid = calculusId;
                 resourceId = integrationId;
@@ -84,7 +87,7 @@ public class TopicResourcesTest extends RestTest {
         // After behavior change: Add the resource again to another topic with primary = false
         // should create a non-primary resource connection
         final var resource2ConnectionPublicId = getId(
-                testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+                testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
                     {
                         topicid = topic2Id;
                         resourceId = integrationId;
@@ -114,7 +117,7 @@ public class TopicResourcesTest extends RestTest {
         final var calculusId = calculus.getPublicId();
         final var integrationId = integration.getPublicId();
 
-        testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 topicid = calculusId;
                 resourceId = integrationId;
@@ -133,7 +136,7 @@ public class TopicResourcesTest extends RestTest {
     public void can_update_topic_resource() throws Exception {
         URI id = save(NodeConnection.create(newTopic(), newResource())).getPublicId();
 
-        testUtils.updateResource("/v1/topic-resources/" + id, new TopicResources.TopicResourcePUT() {
+        testUtils.updateResource("/v1/topic-resources/" + id, new TopicResourcePUT() {
             {
                 primary = true;
             }
@@ -146,7 +149,7 @@ public class TopicResourcesTest extends RestTest {
     public void cannot_unset_primary_topic() throws Exception {
         URI id = save(NodeConnection.create(newTopic(), newResource(), true)).getPublicId();
 
-        testUtils.updateResource("/v1/topic-resources/" + id, new TopicResources.TopicResourcePUT() {
+        testUtils.updateResource("/v1/topic-resources/" + id, new TopicResourcePUT() {
             {
                 primary = false;
             }
@@ -177,7 +180,7 @@ public class TopicResourcesTest extends RestTest {
         save(NodeConnection.create(calculus, integration));
 
         var response = testUtils.getResource("/v1/topic-resources");
-        var topicResources = testUtils.getObject(TopicResources.TopicResourceDTO[].class, response);
+        var topicResources = testUtils.getObject(TopicResourceDTO[].class, response);
 
         assertEquals(2, topicResources.length);
         assertAnyTrue(topicResources, t -> electricity.getPublicId().equals(t.topicid)
@@ -195,7 +198,7 @@ public class TopicResourcesTest extends RestTest {
         var topicResource = save(NodeConnection.create(electricity, alternatingCurrent));
 
         var resource = testUtils.getResource("/v1/topic-resources/" + topicResource.getPublicId());
-        var topicResourceIndexDocument = testUtils.getObject(TopicResources.TopicResourceDTO.class, resource);
+        var topicResourceIndexDocument = testUtils.getObject(TopicResourceDTO.class, resource);
         assertEquals(electricity.getPublicId(), topicResourceIndexDocument.topicid);
         assertEquals(alternatingCurrent.getPublicId(), topicResourceIndexDocument.resourceId);
     }
@@ -208,7 +211,7 @@ public class TopicResourcesTest extends RestTest {
 
         Node graphTheory = builder.node(NodeType.TOPIC, t -> t.name("graph theory"));
 
-        testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 topicid = graphTheory.getPublicId();
                 resourceId = graphs.getPublicId();
@@ -233,14 +236,14 @@ public class TopicResourcesTest extends RestTest {
 
         URI geometrySquares = save(NodeConnection.create(geometry, squares)).getPublicId();
         URI geometryCircles = save(NodeConnection.create(geometry, circles)).getPublicId();
-        testUtils.updateResource("/v1/topic-resources/" + geometryCircles, new TopicResources.TopicResourcePUT() {
+        testUtils.updateResource("/v1/topic-resources/" + geometryCircles, new TopicResourcePUT() {
             {
                 primary = true;
                 id = geometryCircles;
                 rank = 1;
             }
         });
-        testUtils.updateResource("/v1/topic-resources/" + geometrySquares, new TopicResources.TopicResourcePUT() {
+        testUtils.updateResource("/v1/topic-resources/" + geometrySquares, new TopicResourcePUT() {
             {
                 primary = true;
                 id = geometrySquares;
@@ -260,7 +263,7 @@ public class TopicResourcesTest extends RestTest {
                 t -> t.name("elementary maths").resource(r -> r.name("graphs")).resource(r -> r.name("sets")));
 
         var response = testUtils.getResource("/v1/topic-resources");
-        var topicResources = testUtils.getObject(TopicResources.TopicResourceDTO[].class, response);
+        var topicResources = testUtils.getObject(TopicResourceDTO[].class, response);
         assertAllTrue(topicResources, tr -> tr.rank == 0);
     }
 
@@ -270,7 +273,7 @@ public class TopicResourcesTest extends RestTest {
         var squares = builder.node(NodeType.RESOURCE, r -> r.name("Squares").publicId("urn:resource:1"));
         var circles = builder.node(NodeType.RESOURCE, r -> r.name("Circles").publicId("urn:resource:2"));
 
-        testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 primary = true;
                 topicid = geometry.getPublicId();
@@ -279,7 +282,7 @@ public class TopicResourcesTest extends RestTest {
             }
         });
 
-        testUtils.createResource("/v1/topic-resources", new TopicResources.TopicResourcePOST() {
+        testUtils.createResource("/v1/topic-resources", new TopicResourcePOST() {
             {
                 primary = true;
                 topicid = geometry.getPublicId();
@@ -305,7 +308,7 @@ public class TopicResourcesTest extends RestTest {
         var updatedConnection = topicResources.get(topicResources.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/topic-resources/" + updatedConnection.getPublicId().toString(),
-                new TopicResources.TopicResourcePUT() {
+                new TopicResourcePUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -316,7 +319,7 @@ public class TopicResourcesTest extends RestTest {
         // verify that the other connections have been updated
         for (NodeConnection topicResource : topicResources) {
             var response = testUtils.getResource("/v1/topic-resources/" + topicResource.getPublicId().toString());
-            var connectionFromDb = testUtils.getObject(TopicResources.TopicResourceDTO.class, response);
+            var connectionFromDb = testUtils.getObject(TopicResourceDTO.class, response);
             // verify that the other connections have had their rank bumped up 1
             if (!connectionFromDb.id.equals(updatedConnection.getPublicId())) {
                 int oldRank = mappedRanks.get(connectionFromDb.id.toString());
@@ -336,7 +339,7 @@ public class TopicResourcesTest extends RestTest {
         var updatedConnection = topicResources.get(topicResources.size() - 1);
         assertEquals(100, updatedConnection.getRank());
         testUtils.updateResource("/v1/topic-resources/" + updatedConnection.getPublicId().toString(),
-                new TopicResources.TopicResourcePUT() {
+                new TopicResourcePUT() {
                     {
                         primary = true;
                         rank = 1;
@@ -347,7 +350,7 @@ public class TopicResourcesTest extends RestTest {
         // verify that the other connections have been updated
         for (NodeConnection topicResource : topicResources) {
             var response = testUtils.getResource("/v1/topic-resources/" + topicResource.getPublicId().toString());
-            var connectionFromDb = testUtils.getObject(TopicResources.TopicResourceDTO.class, response);
+            var connectionFromDb = testUtils.getObject(TopicResourceDTO.class, response);
             // verify that only the contiguous connections are updated
             if (!connectionFromDb.id.equals(updatedConnection.getPublicId())) {
                 int oldRank = mappedRanks.get(connectionFromDb.id.toString());
@@ -369,7 +372,7 @@ public class TopicResourcesTest extends RestTest {
         var updatedConnection = topicResources.get(topicResources.size() - 1);
         assertEquals(10, updatedConnection.getRank());
         testUtils.updateResource("/v1/topic-resources/" + topicResources.get(9).getPublicId().toString(),
-                new TopicResources.TopicResourcePUT() {
+                new TopicResourcePUT() {
                     {
                         primary = true;
                         rank = 99;
@@ -380,7 +383,7 @@ public class TopicResourcesTest extends RestTest {
         // verify that the other connections are unchanged
         for (NodeConnection topicResource : topicResources) {
             var response = testUtils.getResource("/v1/topic-resources/" + topicResource.getPublicId().toString());
-            var connection = testUtils.getObject(TopicResources.TopicResourceDTO.class, response);
+            var connection = testUtils.getObject(TopicResourceDTO.class, response);
             if (!connection.id.equals(updatedConnection.getPublicId())) {
                 assertEquals(mappedRanks.get(connection.id.toString()).intValue(), connection.rank);
             }
